refactor: move globalConfig.json update logic to a separate file
